+++ lib/fconsole.dart
@@ -2,5 +2,6 @@ library fconsole;
 
 export 'package:fconsole/src/widget/console.dart';
 export 'package:fconsole/src/core/fconsole.dart';
+export 'package:fconsole/src/core/console.dart';
 export 'package:fconsole/src/model/flow.dart';
 export 'package:oktoast/oktoast.dart';