+++ lib/src/core/fconsole.dart
@@ -1,25 +1,42 @@
 import 'dart:async';
 
+import 'package:fconsole/fconsole.dart';
 import 'package:fconsole/src/model/log.dart';
 import 'package:flutter/material.dart';
 
 import '../model/log.dart';
 import 'shake_detector.dart';
 
+typedef ErrHandler = void Function(
+    Zone, ZoneDelegate, Zone, Object, StackTrace);
+
 // TODO: 拦截的log需不需要显示在其他地方?
-void runFConsoleApp(Widget app) {
+void runFConsoleApp(Widget app, {ErrHandler errHandler}) {
   FlutterError.onError = (details) {
     Zone.current.handleUncaughtError(details.exception, details.stack);
   };
 
   var zoneSpecification = ZoneSpecification(
-    print: (Zone self, ZoneDelegate parent, Zone zone, String line) {
-      FConsole.log(line);
+    print: (
+      Zone self,
+      ZoneDelegate parent,
+      Zone zone,
+      String line,
+    ) {
+      FConsole.log(line, noPrint: true);
+      Zone.current.parent?.print(line);
     },
-    handleUncaughtError: (Zone self, ZoneDelegate parent, Zone zone,
-        Object error, StackTrace stackTrace) {
-      // TODO: 堆栈可处理
-      FConsole.error(error);
+    handleUncaughtError: (
+      Zone self,
+      ZoneDelegate parent,
+      Zone zone,
+      Object error,
+      StackTrace stackTrace,
+    ) {
+      // TODO: 堆栈错误可处理
+      FConsole.error(error, noPrint: true);
+      Zone.current.parent?.print(error);
+      errHandler?.call(self, parent, zone, error, stackTrace);
     },
   );
   runZoned(
@@ -33,21 +50,6 @@ class FConsole extends ChangeNotifier {
   ShakeDetector shakeDetector;
 
   ValueNotifier isShow = ValueNotifier(false);
-
-  static FConsole _instance;
-
-  factory FConsole() => _getInstance();
-
-  static FConsole get instance => _getInstance();
-
-  static FConsole _getInstance() {
-    if (_instance == null) {
-      _instance = FConsole._();
-    }
-    return _instance;
-  }
-
-  FConsole._();
 
   void startShakeListener(Function() onShake) {
     stopShakeListener();
@@ -69,16 +71,7 @@ class FConsole extends ChangeNotifier {
   List<Log> verboselog = [];
   int currentLogIndex = 0;
 
-  static log(dynamic log) {
-    if (log != null) {
-      Log lg = Log(log, LogType.log);
-      FConsole.instance.verboselog.add(lg);
-      FConsole.instance.allLog.add(lg);
-      FConsole.instance.notifyListeners();
-    }
-  }
-
-  List<Log> logs(int logType) {
+  List<Log> logListOfType(int logType) {
     if (logType == 0) {
       return allLog;
     }
@@ -91,7 +84,30 @@ class FConsole extends ChangeNotifier {
     return null;
   }
 
-  static error(dynamic error) {
+  static log(dynamic log, {bool noPrint = false}) {
+    // 有这个参数时，是自己捕获的print，不需要再打印
+    if (!noPrint) {
+      print(log);
+      return;
+    }
+    if (FConsole.instance.isShow.value == false) {
+      return;
+    }
+    if (log != null) {
+      Log lg = Log(log, LogType.log);
+      FConsole.instance.verboselog.add(lg);
+      FConsole.instance.allLog.add(lg);
+      FConsole.instance.notifyListeners();
+    }
+  }
+
+  static error(dynamic error, {bool noPrint = false}) {
+    if (!noPrint) {
+      print(error);
+    }
+    if (FConsole.instance.isShow.value == false) {
+      return;
+    }
     if (error != null) {
       Log lg = Log(error, LogType.error);
       FConsole.instance.errorLog.add(lg);
@@ -116,6 +132,22 @@ class FConsole extends ChangeNotifier {
     }
     FConsole.instance.notifyListeners();
   }
+
+  /// 单例
+  static FConsole _instance;
+
+  factory FConsole() => _getInstance();
+
+  static FConsole get instance => _getInstance();
+
+  static FConsole _getInstance() {
+    if (_instance == null) {
+      _instance = FConsole._();
+    }
+    return _instance;
+  }
+
+  FConsole._();
 }
 
 class ConsoleOptions {